Refactor package select body so it can be shared
